+++ src/main/java/core/file/hrf/HRFImport.java
@@ -46,16 +46,20 @@ public class HRFImport {
 
 				// remember path
 				UserParameter.instance().hrfImport_HRFPath = files[i].getParentFile().getAbsolutePath();
+				var teamId = HOVerwaltung.instance().getModel().getBasics().getTeamId();
 				frame.setInformation(getLangStr("StartParse"));
 				homodel = HRFFileParser.parse(files[i]);
-				if ( homodel != null && homodel.getBasics().getTeamId() != HOVerwaltung.instance().getModel().getBasics().getTeamId()){
-					HOLogger.instance().error(getClass(), "hrf file from other team can not be imported: " + homodel.getBasics().getTeamName());
+				var errorMessage = "Importfehler";
+				if ( homodel != null && homodel.getBasics().getTeamId() != teamId){
+					HOLogger.instance().error(getClass(), "Hrf file from other team can not be imported: " + homodel.getBasics().getTeamName());
+					errorMessage = "hrf.import.error.teamid_conflict";
 					homodel = null;
 				}
 
 				if (homodel == null) {
-					frame.setInformation(getLangStr("Importfehler") + " : " + files[i].getName(), InfoPanel.FEHLERFARBE);
-					Helper.showMessage(frame, getLangStr("Importfehler"), getLangStr("Fehler"), JOptionPane.ERROR_MESSAGE);
+					var languageErrorMessage = getLangStr(errorMessage);
+					frame.setInformation(languageErrorMessage + " : " + files[i].getName(), InfoPanel.FEHLERFARBE);
+					Helper.showMessage(frame, languageErrorMessage, getLangStr("Fehler"), JOptionPane.ERROR_MESSAGE);
 				} else {
 					frame.setInformation(getLangStr("HRFSave"));